refactor: integrate @repo/ui components and simplify chat interface
- Replace custom chat components with @repo/ui equivalents where available
- Remove duplicate CSS variables and use @repo/ui globals.css as base
- Fix hardcoded gray backgrounds with semantic design tokens
- Simplify chat interface structure to match apps/chat pattern
- Update dev-server styling with custom dark theme colors
- Remove custom shadcn components in favor of shared UI library
