+++ core/dev-server/src/components/default-catch-boundary.tsx
@@ -24,21 +24,21 @@ export function DefaultCatchBoundary({ error }: ErrorComponentProps) {
           onClick={() => {
             void router.invalidate()
           }}
-          className={`px-2 py-1 bg-gray-600 dark:bg-gray-700 rounded text-white uppercase font-extrabold`}
+          className={`px-2 py-1 bg-primary text-primary-foreground rounded uppercase font-extrabold`}
         >
           Try Again
         </button>
         {isRoot ? (
           <Link
             to="/"
-            className={`px-2 py-1 bg-gray-600 dark:bg-gray-700 rounded text-white uppercase font-extrabold`}
+            className={`px-2 py-1 bg-primary text-primary-foreground rounded uppercase font-extrabold`}
           >
             Home
           </Link>
         ) : (
           <Link
             to="/"
-            className={`px-2 py-1 bg-gray-600 dark:bg-gray-700 rounded text-white uppercase font-extrabold`}
+            className={`px-2 py-1 bg-primary text-primary-foreground rounded uppercase font-extrabold`}
             onClick={(e) => {
               e.preventDefault()
               window.history.back()
