+++ core/dev-server/src/styles/globals.css
@@ -1,136 +1,48 @@
-@import "tailwindcss";
-@import "tw-animate-css";
+/* Import the base UI globals */
+@import "@repo/ui/globals.css";
 
-@custom-variant dark (&:is(.dark *));
-
-@font-face {
-  font-family: "Geist Sans";
-  src: url("/fonts/GeistVF.woff2") format("woff2-variations");
-  font-weight: 100 900;
-  font-style: normal;
-}
-
-@font-face {
-  font-family: "Geist Mono";
-  src: url("/fonts/GeistMonoVF.woff2") format("woff2-variations");
-  font-weight: 100 900;
-  font-style: normal;
-}
-
-@theme inline {
-  --font-family-sans: "Geist Sans", system-ui, sans-serif;
-  --font-family-mono: "Geist Mono", monospace;
-  --radius-sm: calc(var(--radius) - 4px);
-  --radius-md: calc(var(--radius) - 2px);
-  --radius-lg: var(--radius);
-  --radius-xl: calc(var(--radius) + 4px);
-  --color-background: var(--background);
-  --color-foreground: var(--foreground);
-  --color-card: var(--card);
-  --color-card-foreground: var(--card-foreground);
-  --color-popover: var(--popover);
-  --color-popover-foreground: var(--popover-foreground);
-  --color-primary: var(--primary);
-  --color-primary-foreground: var(--primary-foreground);
-  --color-secondary: var(--secondary);
-  --color-secondary-foreground: var(--secondary-foreground);
-  --color-muted: var(--muted);
-  --color-muted-foreground: var(--muted-foreground);
-  --color-accent: var(--accent);
-  --color-accent-foreground: var(--accent-foreground);
-  --color-destructive: var(--destructive);
-  --color-border: var(--border);
-  --color-input: var(--input);
-  --color-ring: var(--ring);
-  --color-chart-1: var(--chart-1);
-  --color-chart-2: var(--chart-2);
-  --color-chart-3: var(--chart-3);
-  --color-chart-4: var(--chart-4);
-  --color-chart-5: var(--chart-5);
-  --color-sidebar: var(--sidebar);
-  --color-sidebar-foreground: var(--sidebar-foreground);
-  --color-sidebar-primary: var(--sidebar-primary);
-  --color-sidebar-primary-foreground: var(--sidebar-primary-foreground);
-  --color-sidebar-accent: var(--sidebar-accent);
-  --color-sidebar-accent-foreground: var(--sidebar-accent-foreground);
-  --color-sidebar-border: var(--sidebar-border);
-  --color-sidebar-ring: var(--sidebar-ring);
-}
-
-:root {
-  --radius: 0.625rem;
-  --background: oklch(1 0 0);
-  --foreground: oklch(0.145 0 0);
-  --card: oklch(1 0 0);
-  --card-foreground: oklch(0.145 0 0);
-  --popover: oklch(1 0 0);
-  --popover-foreground: oklch(0.145 0 0);
-  --primary: oklch(0.205 0 0);
-  --primary-foreground: oklch(0.985 0 0);
-  --secondary: oklch(0.97 0 0);
-  --secondary-foreground: oklch(0.205 0 0);
-  --muted: oklch(0.97 0 0);
-  --muted-foreground: oklch(0.556 0 0);
-  --accent: oklch(0.97 0 0);
-  --accent-foreground: oklch(0.205 0 0);
-  --destructive: oklch(0.577 0.245 27.325);
-  --border: oklch(0.922 0 0);
-  --input: oklch(0.922 0 0);
-  --ring: oklch(0.708 0 0);
-  --chart-1: oklch(0.646 0.222 41.116);
-  --chart-2: oklch(0.6 0.118 184.704);
-  --chart-3: oklch(0.398 0.07 227.392);
-  --chart-4: oklch(0.828 0.189 84.429);
-  --chart-5: oklch(0.769 0.188 70.08);
-  --sidebar: oklch(0.985 0 0);
-  --sidebar-foreground: oklch(0.145 0 0);
-  --sidebar-primary: oklch(0.205 0 0);
-  --sidebar-primary-foreground: oklch(0.985 0 0);
-  --sidebar-accent: oklch(0.97 0 0);
-  --sidebar-accent-foreground: oklch(0.205 0 0);
-  --sidebar-border: oklch(0.922 0 0);
-  --sidebar-ring: oklch(0.708 0 0);
-}
-
+/* 
+ * Custom color overrides for Chat app dark mode only
+ * Light mode uses the default colors from @repo/ui
+ */
 .dark {
-  --background: oklch(0.145 0 0);
-  --foreground: oklch(0.985 0 0);
-  --card: oklch(0.205 0 0);
-  --card-foreground: oklch(0.985 0 0);
-  --popover: oklch(0.205 0 0);
-  --popover-foreground: oklch(0.985 0 0);
-  --primary: oklch(0.922 0 0);
-  --primary-foreground: oklch(0.205 0 0);
-  --secondary: oklch(0.269 0 0);
-  --secondary-foreground: oklch(0.985 0 0);
-  --muted: oklch(0.269 0 0);
-  --muted-foreground: oklch(0.708 0 0);
-  --accent: oklch(0.269 0 0);
-  --accent-foreground: oklch(0.985 0 0);
-  --destructive: oklch(0.704 0.191 22.216);
-  --border: oklch(1 0 0 / 10%);
-  --input: oklch(1 0 0 / 15%);
-  --ring: oklch(0.556 0 0);
-  --chart-1: oklch(0.488 0.243 264.376);
-  --chart-2: oklch(0.696 0.17 162.48);
-  --chart-3: oklch(0.769 0.188 70.08);
-  --chart-4: oklch(0.627 0.265 303.9);
-  --chart-5: oklch(0.645 0.246 16.439);
-  --sidebar: oklch(0.205 0 0);
-  --sidebar-foreground: oklch(0.985 0 0);
-  --sidebar-primary: oklch(0.488 0.243 264.376);
-  --sidebar-primary-foreground: oklch(0.985 0 0);
-  --sidebar-accent: oklch(0.269 0 0);
-  --sidebar-accent-foreground: oklch(0.985 0 0);
-  --sidebar-border: oklch(1 0 0 / 10%);
-  --sidebar-ring: oklch(0.556 0 0);
+	--background: oklch(0.2178 0 0);
+	--foreground: oklch(0.8853 0 0);
+	--card: oklch(0.2435 0 0);
+	--card-foreground: oklch(0.8853 0 0);
+	--popover: oklch(0.2178 0 0); /* Match background for consistency */
+	--popover-foreground: oklch(0.8853 0 0);
+	--primary: oklch(0.7058 0 0);
+	--primary-foreground: oklch(0.2178 0 0);
+	--secondary: oklch(0.3092 0 0);
+	--secondary-foreground: oklch(0.8853 0 0);
+	--muted: oklch(0.285 0 0);
+	--muted-foreground: oklch(0.5999 0 0);
+	--accent: oklch(0.26 0 0); /* Consistent accent color matching sidebar */
+	--accent-foreground: oklch(0.8853 0 0);
+	--destructive: oklch(0.6591 0.153 22.1703);
+	--destructive-foreground: oklch(1 0 0);
+	--border: oklch(0.329 0 0);
+	--input: oklch(0.3092 0 0);
+	--input-bg: oklch(0.24 0 0); /* Very subtle, slightly lighter than background */
+	--ring: oklch(0.7058 0 0);
+	--chart-1: oklch(0.7058 0 0);
+	--chart-2: oklch(0.6714 0.0339 206.3482);
+	--chart-3: oklch(0.5452 0 0);
+	--chart-4: oklch(0.4604 0 0);
+	--chart-5: oklch(0.3715 0 0);
+	--sidebar: oklch(0.2178 0 0); /* Same as background */
+	--sidebar-foreground: oklch(0.8853 0 0);
+	--sidebar-primary: oklch(0.7058 0 0);
+	--sidebar-primary-foreground: oklch(0.2178 0 0);
+	--sidebar-accent: oklch(0.26 0 0); /* Subtle accent, same as input-bg */
+	--sidebar-accent-foreground: oklch(0.8853 0 0);
+	--sidebar-border: oklch(0.329 0 0);
+	--sidebar-ring: oklch(0.7058 0 0);
 }
 
-@layer base {
-  * {
-    @apply border-border outline-ring/50;
-  }
-  body {
-    @apply bg-background text-foreground;
-  }
+/* Additional custom styles specific to the chat app */
+@layer utilities {
+	/* Add any chat-specific utility classes here */
 }
+
